# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: API specification for AOPMapExercise
  description: API specification for AOPMapExercise
  version: 1.0.0
servers:
  - url: http://localhost:8081/
paths:
  /postToDoInfo:
    post:
      tags:
        - createtodo
      summary: Create a todo task
      operationId: postTodoInfo
      requestBody:
        description: Create a Todo task for system
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/CreateToDoRequest'
            examples:
              ToDo1:
                value:
                  authorId: "1"
                  title: "Meet with Doctor Strange"
                  content: "I have a meeting with Doctor Strange at 5pm"
                  priority: "Medium"
                  status: "Done"
      responses:
        '200':
          description: Create Todo successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectCreationSuccessResponse'
  /postNewUsers:
    post:
      tags:
        - createNewUser
      summary: This API is used for creating a new user in the system
      operationId: postUserInfo
      requestBody:
        description: Create a new user for system
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/createUserRequest'
            examples:
              User1:
                value:
                  email: "admin@gmail.com"
                  fullName: "Admin"
      responses:
        '200':
          description: Create Todo successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectCreationSuccessResponse'
              examples:
                SuccessfulMessage:
                  value:
                    responseCode: 200
                    message: "Successfully create a new user information"

  /getUserList:
    get:
      tags:
        - getUserList
      summary: Get the list of user in this system
      operationId: GetListUser
      parameters:
        - in: query
          name: order
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
        - in: query
          name: field
          schema:
            type: string
        - in: query
          name: page
          description: page number, start at page 0, each page will return 10 item
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Get the list of user in system
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'

  /getTaskList:
    get:
      tags:
        - getTaskList
      summary: Get the list of task in this system
      operationId: GetTaskList
      parameters:
        - in: query
          name: order
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
        - in: query
          name: field
          schema:
            type: string
        - in: query
          name: page
          description: page number, start at page 0, each page will return 10 item
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Get the list of task in system
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'

components:
  schemas:
    CreateToDoRequest:
      type: object
      required:
        - authorId
        - title
        - content
        - priority
        - status
      properties:
        authorId:
          type: string
        title:
          type: string
        content:
          type: string
        priority:
          type: string
          enum:
            - Low
            - Medium
            - High
        status:
          type: string
          enum:
            - Done
            - NotDone
    createUserRequest:
      type: object
      required:
        - email
        - fullName
      properties:
        email:
          type: string
          format: email
          pattern: '\S+@\S+\.\S+'
        fullName:
          type: string
    UserListResponse:
      type: object
      properties:
        currentPage:
          type: integer
        totalPage:
          type: integer
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserResponseModel'
    UserResponseModel:
      type: object
      properties:
        name:
          type: string
        email:
          type: string

    TaskListResponse:
      type: object
      properties:
        currentPage:
          type: integer
        totalPage:
          type: integer
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponseModel'
    TaskResponseModel:
      type: object
      properties:
        authorName:
          type: string
        title:
          type: string
        content:
          type: string
        priority:
          type: string
        status:
          type: string
    ObjectCreationSuccessResponse:
      type: object
      properties:
        responseCode:
          type: integer
        message:
          type: string
    ObjectUpdateSuccessResponse:
      type: object
      properties:
        responseCode:
          type: integer
        message:
          type: string
  examples:
    InternalServerError:
      value:
        errorCode: 1500
        faults:
          -:
          userMessage: There're some error when processing this request. Please retry again or contact the site administrator for supporting
          internalMessage: Internal server error
          code: 1500
          errorCode: 1500